wip: integrate BE3 services and slot assignment

+++ parking-backend/src/main/java/com/swp391/parking/service/VehicleService.java
@@ -5,10 +5,10 @@ import com.swp391.parking.entity.Vehicle;
 import java.util.List;
 
 public interface VehicleService {
-    List<Vehicle> getByUser(Integer userId);
+    List<Vehicle> getByUser(Long userId);
     Vehicle getById(Long id);
     Vehicle getByLicensePlate(String licensePlate); // Staff dùng khi xe vào/ra
-    Vehicle create(Integer userId, VehicleRequest req);
+    Vehicle create(Long userId, VehicleRequest req);
     Vehicle update(Long id, VehicleRequest req);
     void deactivate(Long id);
 }
